+++ src/MusicLibraryApi.Client/Interfaces/IDiscsMutation.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System.IO;
+using System.Threading;
 using System.Threading.Tasks;
 using MusicLibraryApi.Client.Contracts.Discs;
 
@@ -7,5 +8,7 @@ namespace MusicLibraryApi.Client.Interfaces
 	public interface IDiscsMutation
 	{
 		Task<int> CreateDisc(InputDiscData discData, CancellationToken cancellationToken);
+
+		Task<int> CreateDisc(InputDiscData discData, Stream coverContent, CancellationToken cancellationToken);
 	}
 }