+++ src/MusicLibraryApi.Client/ReleaseNotes.md
@@ -1,4 +1,10 @@
-﻿**3.0.0** / 2019.12.16**
+﻿**4.1.0** / 2020.01.12**
+* Added storing of disc cover images.
+
+**4.0.0** / 2020.01.07**
+* Added storing of song content.
+
+**3.0.0** / 2019.12.16**
 * Fixed client settings setup.
 
 **2.0.0** / 2019.12.16**